[BorderDOFExchanger] Do not include constrained matrix entries
Matrix entries involving constrained DOFs (at least Dirichlet constraints) are already
correctly initialized by the constraints code with the corresponding unit vector
and should not be summed up. Well, it really shouldn't matter because due to the
residual formulation, the right hand side for a Dirichlet-constrained DOF is
0 anyway, but better be on the safe side here...

On the other hand, it *is* important to make sure the extended pattern retains
its symmetricity properties - otherwise AMG might break in a very confusing
manner.

In order to efficiently test for constraints, the BorderDOFExchanger needs
access to the local indices and the IndexCaches, so I seized the opportunity
and moved all of the looping and setup logic still in the pattern engine to
the BorderDOFExchanger. That way, alternative GridOperator with custom
pattern engines won't have to reinvent the wheel...