[Ordering] Add new mixin PartitionInfoProvider for implementing PartitionType query interface
This patch adds a new class PartitionInfoProvider that will be used to add a query
interface for contained grid PartitionTypes to all orderings. As this information
is completely independent of any template parameters, we can easily centralize the
implementation in a single location. The only thing the orderings will need to do
is update the contained information.