[Ordering] Improve tags for leaf orderings to also contain information about contained grid PartitionTypes
The new ordering infrastructure supports GridFunctionSpaces on grid partitions other
than All_Partition, but unfortunately, we have no way of extracting the exact nature
of the covered partition from the underlying GridView (I haven't found one, at least...).

With this patch, the user can provide this information via the ordering tag on the leaf
spaces (on non-leaf spaces, we can simply combine the information from the child spaces).

For now, the default tag will assume a GridFunctionSpace on All_Partition, while the
NonOverlappingLeafOrderingTag works on GridViews with InteriorBorder_Partition.

Also improved the documentation a little bit. :-)