[GFSDataHandle] Make data handles work correctly for GridFunctionSpaces on InteriorBorder_Partition
The GFSDataHandle implementation has to be extended to make it work for
GridFunctionSpaces defined on the InteriorBorder_Pattern of a grid that possesses
overlap and / or ghost entities.

The current implementation of the GFSDataHandle uses three levels of components
for the gather / scatter logic:

* GFSDataHandle itself is responsible for the communication structure (which is
  also outsourced to a functor), for identifying the DOFs associated with a
  grid entity and for loading (and, if requested, saving) the DOF values.

* One of several intermediate GatherScatter implementations takes care of iterating
  over the DOFs and invoking a per-DOF GatherScatter functor for each DOF, passing
  an implementation-dependent set of parameters.

* One of a wide array of per-DOF GatherScatter implementations doing things like
  copying, adding, subtracting, minimizing... DOF values.

For nonoverlapping GridFunctionSpaces and depending on the chosen communication
interface, we can end up with a situtation where a process receives data for a
grid entity without any DOFs attached to it, e.g. when communicating on the
InteriorBorder_All interface and receiving an entity which is locally a ghost.

In this case, the intermediate GatherScatter functor should simply drain the
communication buffer and return control to the grid communication routine. Here,
this implemented in a generic fashion by querying the GridFunctionSpace whether
the PartitionType of the current entity is part of the GridFunctionSpace. We also
check whether the amount of received data matches the number of DOFs in the case
where the entity is contained and make sure there really are no DOFs present
otherwise.

Note that while with this patch, specifying an overly large receive partition will not
break the call to communicate(), it is still better to correctly specify the
communication interface to avoid unnecessary gather operations and data traffic.
That said, common code for overlapping and non-overlapping grids might sometimes
require the larger receive partition.