[DataHandle] Improve DataHandles for communication structure setup
The DataHandles used for obtaining the information required to set up parallel
solvers are a little bit of a mess - some of them are bundled with the GFSDataHandle,
but most of them are *nested classes* inside ParallelISTLHelper, which is really
unfortunate as their functionality will probably also be required for writing other
parallel LA backends (PETSc anyone?). Moreover, they do not really make use of the
common infrastructure of the GFSDataHandle and only work for simple matrix structures.

This patch gathers all those DataHandles in genericdatahandle.hh and converts most of them
to use the GFSDataHandle. In particular, we

* merge the ghost data handles from genericdatahandle.hh and the ParallelISTLHelper and
  minimize the involved communication volume by only transferring a single piece of data
  per entity, irrespective of the number of DOFs involed.

* replace the PartitionDataHandle with a new DisjointPartitioningDataHandle, which also
  optimizes the communication volume.

* move the GFSNeighborHandle to genericdatahandle.hh. This data handle will construct the
  set of neighboring GFS ranks which share entities that the GFS attaches DOFs to.

* move the SharedDOFDataHandle to genericdatahandle. This data handle will flag any DOFs
  that exist on more than one processor.

All of those new implementations only transmit a single piece of data per entity, as their
operation only requires per-entity (and not per-DOF) information on the scatter side.

The patch also improves documentation a bit and cleans up the includes.