[ISTL] Add BlockMatrixDiagonal for representing (and inverting) diagonal of ISTL block matrices
The new BlockMatrixDiagonal can be used to extract the block diagonal of a possibly nested
ISTL matrix. The container supports block-wise inversion and matrix-vector multiplication and
provides a data handle to make the diagonal consistent on a nonoverlapping grid decomposition. Its
main use for now will be as a building block for a parallel, nonoverlapping Jacobi preconditioner.