[Ordering] Move ordering infrastructure to separate directory
As the ordering infrastructure has developed into something of an independent
component within PDELab (and the amount of code involved is rather substantial),
the gridfunctionspace directory has become really crowded and difficult to grok.

This patch moves all of the ordering-related files in gridfunctionspace/ to the
new directory ordering/ and updates the relevant includes. I've also taken
advantage of the big change to resolve some smaller problems:

* orderingdynamicbase.hh and localorderingdynamicbase.hh had the "dynamic" stripped
  out of their names - it's redundant and now the file names actually match the
  names of the contained classes.

* Improved some (but not all) include order problems, so that people should normally
  fine just including the header for the gridfunctionspace they use.

* renamed orderingutility.hh to utility.hh, as it can now be distinguished by the
  directory.