[SubOrdering] New class SubOrdering for representing subtree views of an Ordering
In order to recreate the functionality of the GridFunctionSubSpace, we need a way
to represent a subset of an Ordering, as all index-processing has been moved from
the GridFunctionSpace to the Orderings.
A SubOrdering represents a mapping from the DOFIndices associated with a subtree
of a GridFunctionSpace (and which appear to be rooted in that subtree, i.e. they
don't contain the portion of the tree from the root of the subtree up to the real
root space) to the ContainerIndices of the appropiately embedded DOFIndices w.r.t.
the root space.
There are some important differences in usage when compared to the existing
GridFunctionSubSpace:

- SubOrderings are not limited to subtrees that are direct children of the root ordering;
  They take a TypeTree::TreePath template argument describing the path from the root
  node to the node the user is interested in.

- SubOrderings may *not* be nested - doing so can lead to anything from compilation
  errors to subtle run-time problems. There is no protection in place to keep users
  from nesting SubOrderings, but I don't think that should really pose a problem, as
  users are expected to only use SubOrderings through a new implementation of the
  GridFunctionSubSpace, which will in turn take care of the issue.

- The TreePath to an ordering corresponding to a subspace of a GridFunctionSpace will
  usually differ from the TreePath leading to that subspace. Again, this will be taken
  care of automatically by the new GridFunctionSubSpace, which will deduce the correct
  TreePath for the SubOrdering.