[ISTL] Move vector backend descriptor to istl/tags.hh
As the StdVectorBackend is broken (and not a good default anyway), it
would be nice to switch to a flat ISTLVectorBackend by default. Unfortunately,
that requires the template ISTLVectorBackend to be available in gridfunctionspace.hh,
which is rather heavywheight to include by default. So this patch cuts it out
from istlvectorbackend.hh and moves it to the much smaller istl/tags.hh.