[shared_ptr] Move ensure_shared_ptr to more reasonable header and fix std::shared_ptr dependency
ensure_shared_ptr() is not really related to the implementation of
FilteredCompositeNode, so move it to a utility header. As it might be a nice
tool to have in other situtations as well, it shouldn't be hidden away in the
TypeTree utilities, so we move it to a new general utility header in common/.

The current implementation of ensure_shared_ptr() also has the problem of only
working with std::shared_ptr. Fixed by removing the namespace qualification.