[FilteredCompositeNode] Remove unnecessary usage of ensure_shared_ptr()
ensure_shared_ptr() should be used when the exact type of a parameter is not
known, but in this situation, the function very much ensures that we can only
ever get a reference. In that case we assume that the user will take care of
storage management and wrap it with stackobject_to_shared_ptr().