[LocalOperator] Don't depend on specific type of local matrix
Ages ago, we dropped the guarantee that the matrix object passed
to the assembly methods of the local operator will be of any
particular type. Looks like we forgot interface.hh and sum.hh
when adapting to that change, as well as removing obsolete include
statements from several more files.