[GridOperatorSpace] Remove GridOperatorSpace and variants and move utility functions away from gridoperatorspace/ directory
This patch removes the deprecated GridOperatorSpace and its variants
InstationaryGridOperatorSpace and MultiStepGridOperatorSpace. Keeping
them around as deprecated is not useful anymore, as they will not even
compile with the new LA backends.

In order to avoid confusion w.r.t. to the similarly-named GridOperator,
we also move all utilities from the old implementations that are still
in use away from the gridoperatorspace/ directory. This mostly concerns
localmatrix.hh, which is simply moved to gridoperator/common, and
gridoperatorspaceutilities.hh, which was some kind of dumping ground for
all kinds of stuff. The contents of that file was mostly removed, the
remainder can now be found in:

- gridoperator/common/assemblerutilities.hh for the LocalSparsityPattern
- localoperator/callswitch.hh for the TMP-based call guards protecting the
  LocalOperator assembly methods

Stuff that's gone now (it wasn't anywhere in PDELab anymore0=):

- std::vector-based matrix backend, doesn't work anymore
- MultiGeomUniqueIDMapper, switch to GlobalGeometryTypeIndex if you still
  use it

As both of these files were included in most example LocalOperators, many
users will also have those includes in their custom operators, so let's
keep them for now and forward to the new headers along with a warning.
Note, however, that neither of those files is required for standard
LocalOperators anymore.

While checking whether the GridOperator implementation passes headercheck,
I noticed some problems which are also fixed by this patch.

Finally, I had to completely reindent localmatrix.hh, as it was full of
tabs.