[LFSIndexCache] Port DOFIndexViewIterator to iterator facade from dune-common to remove Boost dependency
I originally implemented the DOFIndexIndexViewIterator using an iterator_adaptor from Boost
because that seemed the easiest way to go - I only had to provide the changed value / reference
type and override the dereference() method.
While that was fine during development, we do not want to introduce a hard dependency on Boost,
so this patch replaces the Boost component with the RandomAccessIteratorFacade from dune-common.