[LFSIndexCache] Work around the missing support for value types as reference in dune-common iterator facades
The DOFIndexViewIterator returns temporary views on the underlying index when dereferenced,
which requires special support by the iterator facade for operator->(), because we cannot
simply form a pointer to a temporary.

This patch fixes the problem using a trick I found in Boost's iterator facade library: Return
a proxy object that stores a copy of the temporary returned by dereference(), which works because
the standard mandates that the compiler automatically evaluates operator->() a second time on
the proxy object.

Another awkward reimplementation of something Boost solved generically ages ago...