[Tests] Add rudimentary unit test for Raviart-Thomas FEMs
testrtfem is a new unit test that tries to instantiate all implemented variants
of Raviart-Thomas FEMs and runs some basic tests on those FEMs. For now it only
tries to find a finite element for a given cell of the grid, but this should be
extended to verify the correct orientation of the elements w.r.t. their neighbors.