[FEM] Add RaviartThomasLocalFiniteElementMap as frontend for the different implementations
With the newly contributed Raviart-Thomas FEM implementations, the number of implemented
variants has gotten quite large, and taken together with the awkward naming scheme of both
the headers and the FEM classes, changing the FEM becomes error-prone.

In order to work around this problem, we introduce a single frontend class called
RaviartThomasLocalFiniteElementMap that automatically selects the appropriate implementation
based on the GridView and the selected element order. As the type of the FEM depends on the
geometry type of the grid cells, we need to somehow guess that information statically. This
is currently implemented by taking the topologyId of the grid cells as the final template
parameter, which is defaulted to the value from the grid capability hasSingleGeometryType.
If that capability is not set (e.g. for UG), the user will have to explicitly provide the
parameter, but can still use the frontend.