[Tests] Port Raviart-Thomas tests to new, unified FEM interface
All tests based on RT FEMs now use the RaviartThomasLocalFiniteElementMap instead
of dimension / topology / order specific classes. Note that in case of testrt0 and
testrt02dgridfunctionspace, it is necessary to explicitly provide the topology id
of the grid cells because those tests use UG, which may be hybrid and thus does
not export a default topology type at compile time.