[FEM] Improve element geometry handling in front end FEMs for RT and BDM
My initial patch exposed users to the ugly topologyId. We really don't want users to
have to wrangle those GenericGeometry internals, so this patch shuffles the template
parameters around to use GeometryType::BasicType instead. It relies on the newly introduced
BasicTypeFromDimensionAndTopologyId to guess the BasicType from the topologyId exported
by the hasSingleGeometryType capability.