[ParallelHelper] Fix warning about shifting by negative value in PartitionSelector
PartitionSelector statically creates a bitmask from a variable number of shift values
and uses the value -1 to identify an unused shift slot. The code correctly identifies
those ignored values and avoids shifting by them (which would lead to undefined behavior),
but gcc does not recognize that guard.

This patch avoids the problem by moving the switch between shifting and not shifting into
a template specialization.