[ISTL] Add tuple-based implementation of iterator extraction TMP for vector_iterator
My original implementation of the TMP building the stack of iterator types for
the vector_iterator relies on variadic templates to iteratively extend the list
of types.

This patch adds a fallback implementation that stores the types in a tuple and
uses the PushBackTuple helper to append to the tuple. The new implementation is
only picked if the compiler does not support variadic templates.