[Ordering] Offsets in lexicographic orderings should be based on block counts, not flat size
When testing a blocking configuration with a LexicographicOrdering on top of identically blocked
GridViewOrderings, Felix noticed that OrderingBase ignores the blocking structure of the children
and calculates per-child offsets based on the scalar size of each child.

This patch fixes the problem by tracking the scalar size and the block count of the children separately.
It also changes the names of the associated functions: offset(), which used to return the flattened offset,
is now called sizeOffset(). The block offset can be obtained with blockOffset().