[GridFunctions] Added a new gridfunction VectorDiscreteGridFunctionGradient and correspoding test
testutilities.cc now contains another function testing the gridfunctions, which has to be enhanced.

!!! WARNING: !!!
I had to replace the lines

---
const typename Traits::ElementType::Geometry::JacobianInverseTransposed&
          JgeoIT = e.geometry().jacobianInverseTransposed(x);
---

by

---
const typename Traits::ElementType::Geometry::JacobianInverseTransposed
          JgeoIT = e.geometry().jacobianInverseTransposed(x);
---

(do not use reference) in gridfunctionutilities.hh in order to make it work for gcc-4.7.2.
Otherwise JgeoIT is (for YaspGrid) not valid and contains zero or unitialized values! Under gcc-4.6.3, this does apparently not happen. Steffen is going to file a bug about that for the Dune developers.