[HangingNodeManager] Remove vertex mapper and avoid instantiation of vertex entities
We don't need an elaborate mapper for vertices, there is only a single GeometryType
for them anyway; just use the plain IndexSet instead. The IndexSet provides the
additional benefit of being able to calculate subindices directly without first
instantiating the associated vertex subentity, which can be a rather expensive operation
on some grids.