[Assembler] Insert diagonal rows for *all* constraints
I've been misled by the name of the method handle_dirichlet_constraints(), so that when
porting it to the new infrastructure, I put in a condition that would only insert diagonal
rows into rows that really had a Dirichlet constraint.

Unfortunately, the original implementation does instead clear all constrained rows,
independent of the constraints type, and the new version doesn't work correctly for
hanging node constraints.

Revert to the old behavior.