[boilerplate] Make boilerplate CG constraints classes compatible with new NonoverlappingConformingConstraints
In order to make the NonoverlappingConformingConstraints work with the new infrastructure,
we had to modify it to be templated on the GridView and to take the GridView as a constructor
parameter.

Unfortunately, the GridView is not available in the current boilerplate constraints wrapper
classes. This patch solves the problem by adding the GridView as an additional template parameter
and by providing an additional constructor that takes a GridView instance as parameter. For
backward compatibility, the type of the GridView defaults to the leaf grid view, so that
existing code should continue to work out of the box.

I also renamed the Grid template parameter of the constraints wrappers from "T" to "Grid" for
clarity.