[Constraints] Add unconstrained specializations for all constraints-based DOF manipulation functions
This patch specializes all remaining DOF manipulation functions that act on
constraints on EmptyTransformation to make sure that those functions can be called
for unconstrained spaces without generating an error.

In an effort to clean up documentation a little, all specializations have also
been hidden from Doxygen (the fact that we specialize here is a mere implementation
detail).