[Constraints] Fix constraints assembly for DOFs with both Dirichlet and non-Dirichlet constraints
Adrian found an error in one of the 3D hanging nodes examples in poisson.cc (part of
the Howto), where hanging nodes on Dirichlet-constrained boundaries would lead to wrong
results. The root cause of the problem turned out to be the handling of DOFs that receive
constraints from multiple sources: In the old code, the last constraint to get added would
always win, and in the case of hanging nodes, that turned out to be the boundary constraint,
doing the right thing in this case by applying the Dirichlet constraint only. The new code
reshuffled things a little, causing hanging nodes on the boundary to be hanging-node constrained
only, which amounts to a Neumann 0 boundary.

This patch fixes the problem once and for all by making sure that Dirichlet-style constraints
always get precedence over all other types of constraints, independent of the order in which
they are assembled.