[Constraints] Add isNeumann() methods to default ConstraintsParameter classes
In code using MultiDomain, !isDirichlet() does not always imply a Neumann boundary
condition. So simple operators that do not have a dedicated parameter interface should
use the isNeumann() method of the ConstraintsParameter class to decide whether or not
to assemble Neumann boundaries for a given intersection instead of using !isDirichlet().

Unfortunately, this is a change that breaks backwards compatibility, but this patch
at least fixes all tests based on the Poisson operator as well as adding a correct
isNeumann() method to the default ConstraintsParameter implementations.