[GridViewOrdering] Fix propagation of grid partition information for blocked GridViewOrdering
The GridViewOrdering implementations need to manually copy the partition information
from the local ordering, as the default partition set merging code in the OrderingBase
constructor doesn't run for GridViewOrderings (their children are LocalOrderings, not
Orderings).

This copy step was done correctly for the more common case of the LeafGridViewOrdering,
but forgotten in the GridViewOrdering for blocked spaces.