[ISTLMatrixBackend] Move support infrastructure into istl namespace and hide it from Doxygen
This is a first step towards cleaning up the mess in the backend headers. For some reason,
moving the tag-based helper functions into the new namespace also exposed a number of ordering
problems that the compiler accepted before, so I had to shuffle the order of the functions within
the file a little.