[ISTLVectorBackend] Move support infrastructure into istl namespace and hide it from Doxygen
This is a first step towards cleaning up the mess in the backend headers. Like in the matrix
header, this exposed a number of code ordering problems, so some functions have moved up / down
within the file.

The patch also removes all the bits and pieces of the monstrous TMP that deduces the vector block
structure from the Doxygen docu and adds a little bit of documentation for the raw() helpers.