[ISTLMatrixBackend] Store ISTL object in shared pointer to allow reattaching to different matrices
As outlined in a recent discussion on the mailing list, being able to have a single wrapper
object operate on multiple ISTL matrices is something we want, just like it is already possible
for the vector backend.

So this patch changes the ISTLMatrixContainer to store the ISTL matrix in a shared_ptr and adds
attach() and detach() methods equivalent to the ones in the ISTLBlockVectorContainer.

Beware: The recently added default constructor does NOT allocate an underlying matrix object,
you will have to attach one before using the wrapper!