[ISTL] Add new tags and associated tagged constructors to allow creation of unattached vector / matrix wrappers
This patch adds a new header file with tag structs that can be used to control whether the wrappers
will create a new ISTL object during creation.

As it is now possible to have wrappers that do not contain a valid ISTL object, the assignment operators
need to be a bit more careful and distinguish between having an attached object (call the assignment
operator for the ISTL object) and not being attached (allocate a new ISTL container and invoke the copy
constructor).