[ISTLVectorBackend] Move support infrastructure into separate file istl/vectorhelpers.hh
This patch moves all of the infrastructure required to make the ISTLVectorBackend tick,
but that is not part of its interface, to a separate file.

It also adds forward declarations of both the vector and the matrix wrappers to tags.hh
along with the forward declarations of several ISTL types to avoid having to include
one backend from the other one.