[Backend] Add dense backend with vector and matrix wrappers
This patch adds a very simple backend that provides scalar dense matrices
and vectors. Both wrappers are templated on their storage engine, which
has to be a std::vector-like container in that it provides similar constructors,
element access via [], a size() method and standard iterators. The matrix
is currently in row-major format.

Right now, you need a recent C++11-compatible compiler with support for template
aliases and std::bind to use this backend.