[DenseBackend] Make data structures compatible with basic ISTL solvers
Right now, the dense backend cannot really be tested because there are
no solvers. Fortunately, ISTL places only a very small amount of
requirements on the containers within its solvers. So this patch makes
it possible to use ISTL solvers (with a Richardson preconditioner) by
adding the following three features:

- Export a field_type from VectorContainer
- Implement MatrixContainer::mv()
- Implement MatrixContainer::usmv()