[Backend] Extract uncached matrix views to an external class
Like the vector views, the matrix views are not very container-specific
as long as they don't attempt to perform any clever caching and just
forward most operations to the global container, so this patch extracts
them into external utility classes. Along the way, we also fix a bug
where some modifier methods were erroneously declared const and also
finally provide a const view for matrices that was missing until now.

Just as for the vector views, the new components deprecate global_container()
in favor of container().