[Backend] Deprecate broken and unused file vectorutilities.hh
There are a couple of helper functions in there that only work for
ISTL, and even for ISTL they have been broken since the rewrite.
As there haven't been any complaints about that yet and because
that file isn't included anywhere in either PDELab or the Howto,
deprecating it shouldn't really hurt anybody.