[GridFunctionSpace] Store OrderingTag instance in GridFunctionSpace
Just as with the Backend, there are scenarios where it is convenient to pass not
only compile-time, but also run-time data to an Ordering. So this patch changes
the GridFunctionSpaces to store an instance of the OrderingTag and make it accessible
through orderingTag(). It also adds new constructors to provide a custom OrderingTag
instance to the GridFunctionSpace, but also makes sure not to break existing code by
default-instantiating the OrderingTag if it is not passed explicitly.