[Ordering] Move update_direct_children visitor to utility.hh and into ordering namespace
The functionality of update_direct_children is required by all non-local composite orderings,
so move them to utility.hh. As it is a utility component, it's probably also a good idea to
move it into a subnamespace to reduce the risk of name clashes and to hide it from DOXYGEN,
so do that as well.