[Ordering] Move transformation setup to dedicated header file
The main benefits of this move are less clutter in the ordering files and having all of the
transformation registrations and descriptor declarations happen only once instead of once per
ordering. As they are only declarations, they are allowed to appear more than once, but that
just cries out for problems if they ever need to be changed.