[Typetraits] Add helper function declptr()
The function template declptr() fulfils a role similar to the C++11 function declval().
While the latter returns an rvalue reference to the given type, declptr returns a pointer,
which carries less risk of an implicit instantiation of the parameter type and is backwards
compatible to C++03.

Mostly useful in TypeTree context.