[Ordering] Fix spurious memory corruption when copying / moving Orderings with delegates
When copying or moving an Ordering with a delegate, the default constructors
incorrectly set the delegate pointer to the address of the child object within
the source object. That can create problems when the source object is later
freed and the memory reused. In that case, the delegate contains invalid data,
and as the only call ever performed through the interface is virtual, there
is a high likelihood of executing arbitrary garbage after the vtable pointer
of the delegate got thrashed.

I really don't know how nobody has stumbled over this one before...