[GridFunctionSpace/Ordering] Feed size information from Orderings to GridFunctionSpaces and introduce GFS base class
Disclaimer: This patch has gotten a little big, but it touches too many intertwined
pieces to separate any further...

One important feature of the way the Orderings are created from the GridFunctionSpaces
now is the fact that there should only be a single Ordering tree created from the
root of the GFS tree, reducing the number of template instantiations and thus reducing
compile times. Additionally, this scheme also reduces the number of grid traversals during
function space setup for composite spaces. Unfortunately, these advantage are lost as
soon as the size() method is called on one of the child spaces - that creates a separate
Ordering tree for that child space and immediately traverses the grid to calculate the
size information. The main problem here is the lack of feedback from the Ordering tree
to the GFS tree.

In order to resolve this issue, this patch introduces a very simple, commen base class
for all GridFunctionSpaces that holds size information like the size of the space, the
block count, the global size etc. This information is then updated from the Ordering
tree whenever the Orderings are updated through the GridFunctionSpace interface.

As this change required touching all the relevant methods on the GridFunctionSpaces, I
also went ahead and consolidated as many methods as possible into a new class called
GridFunctionSpaceBase, eliminating quite a bit of duplicate code in the process.

WARNING: This patch is not backwards-compatible in all cases! It is not possible anymore
to call size() or similar global methods directly after constructing a space - doing so
will raise an exception! To avoid the problem call ordering() or update() before size()
et. al.

Some smaller gems in this patch:

- GridFunctionSpace lost a bunch of old variables still left from ancient times
  (before there were Orderings).

- GridFunctionSpaces now track whether they are at the root of the tree.