[Ordering] Add new decorator mechanism to inject nodes with additional functionality into the Ordering tree
The decorator mechanism makes it possible to wrap ordering tags in decorator
tags that will cause additional Ordering nodes to be inserted into the Ordering
tree directly above the decorated node.

The mechanism automatically takes care of making the new Ordering and tag compatible
 with all existing kinds of Orderings and allows for the stacking of multiple decorator
nodes on a single GridFunctionSpace OrderingTag.