[VectorGridFunctionSpace] Fix construction sequence to avoid segfault related to now handling of function space tree
The GridFunctionSpace base class now assumes that the underlying tree is correctly
initialized with valid objects by the time its constructor is called because it has
to propagate information down that tree.
Unfortunately, the VectorGridFunctionSpace postpones the creation of the per-component
child spaces until after that constructor call, causing a segfault.

This patch fixes the problem by moving the population of the storage array to a static
helper function that gets called earlier in the sequence.