[LocalFunctionSpace] Export subspace depth for LFSIndexCache
Due to the recent changes to the way subspaces work, the LFSIndexCache
needs to know how the length of the path from the subspace to the root
of the underlying tree because it needs to remove that portion of the
treeindex before invoking the orderings for the subtree (those expect
to get a shorter DOFIndex than what we have right now, as we have cut
off the upper part of the tree in the subspace).

This patch adds a new method subSpaceLength(), which returns 0 for
regular LocalFunctionSpaces and a positive value for the root
LocalFunctionSpace of the LocalFunctionSpace tree generated from a
GridFunctionSubSpace.