[localassembler/patternengine] move the definition of GlobalDOFIndex and BorderPattern to the BorderDOFExchanger class
We want to get rid of any GridView requirements in the interface. If we operate on different
grids the assembler doesn't have a unique BorderDOFExchanger, no unique GlobalDOFIndex
and no unique BorderPattern. Thus we delay the dfinition as far as possible. Only the
pattern_engine knows about the BorderDOFExchanger (and actually even this class shouldn't
know about it!).