[LFSIndexCache] Replace C arrays that might be of length 0 with array class
The standard says that C-style arrays of length 0 are not supported,
which triggers a warning in LFSIndexCache if the underlying LFS has
no children.

The std::array container does support zero-length arrays, however, so
this patch just replaces the C array with a Dune::array (which is either
a std::array or a std::tr1::array on all platforms supported by PDELab
and should thus have the correct specialization for zero length).

Thanks to Christoph Grüninger for the tip-off.