[ordering] add paranoiy checkt to ensure that the geometrytype is compatible with what we precomputed
Some FEM implementations use hard-coded GeometryTypes, this can lead to the ugly siutation,
that we precompute certain offsets for these GeometryTypes, but the grid acutally yields other
types. In this situation the container indices are wrong. This additoinal check makes sure that
we con't accidentially run into this problem. On the long run we should fix the FEM implementations