[test] update testordering to trigger a bug, I've been hunting for several days
VariableMonomFEM on triangular meshes lead to ``unexpected'' results. One could also say
it totally fails. This worked for older versions of PDELab, but since PDELab got a clever FEM and
can reduce the memory footprint to compute DOF indices, we run in problems, when the
GeometryType doesn't fit to that of the mesh.