[ISTLBackend][Bugfix] Fix handling of static_blocking with block_size == 1
When the user requests blocking in the ISTLVectorBackend, we honor this
request even if it is redundant, i.e. the blocking is static and the
block size is 1.
This behavior creates a problem in the ISTL backends, which expect the
redundant index information at that level to be elided, and keeping it
in here breaks vector and matrix accesses.

This patch works around that problem by simply overriding the user and
turning off blocking internally.