[BugFix][GridView] Don't *ever* store a reference to a GridView beyond the current scope
If you are given a GridView, it tends to be a temporary, in line with
their concept as cheap throwaway objects. It is thus not a good idea to
store a reference to such an object beyond the current scope, in
particular as a member variable - and yes, everyone including me is
guilty of this... :-(

After spending half a day hunting another stupid bug because of a
reference to a destroyed temporary (that annoyingly only showed up in
optimized mode on a recent GCC because only GCC reuses stack space by
now and only does so in optimized mode), I took the rather extreme
measure of grepping the complete PDELab source code for occurences of
long-living references and purged every single one of them.

So this patch turns all long-living const references to GridViews into
non-const objects. I directly went ahead and made them non-const so we
will be able to deal with the fact that we will have to update them once
grids start handing out GridViews that only stay valid until the next
grid modification.